--- __needs/index.js
@@ -1,11 +1,0 @@
-module.exports = {
-  /* helpers */
-  Server : require ('./Server'),
-  Router : require ('./Router'),
-  /* project map */
-  // data   : require ('../data'),
-  // models : require ('../models'),
-  // routes : require ('../routes'),
-  // server : require ('../server'),
-  // tools  : require ('../tools'),
-}
